[refactor] remove all the non-simtest structs. Rename to AuthorityNode.
