+++ consensus/core/src/swarm/mod.rs
@@ -1,11 +1,5 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
-pub mod node;
-
 #[cfg(msim)]
-#[path = "./container-sim.rs"]
-mod container;
-
-#[cfg(not(msim))]
-#[path = "./container.rs"]
-mod container;
+#[path = "./node.rs"]
+pub mod node;
