+++ consensus/core/src/swarm/node.rs
@@ -1,16 +1,23 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
-use super::container::AuthorityNodeContainer;
+
+use arc_swap::ArcSwapOption;
+use mysten_metrics::monitored_mpsc::UnboundedReceiver;
+use std::{
+    net::{IpAddr, SocketAddr},
+    sync::Arc,
+    time::Duration,
+};
+use tracing::{info, trace};
+
 use anyhow::Result;
 use consensus_config::{AuthorityIndex, Committee, NetworkKeyPair, Parameters, ProtocolKeyPair};
-use mysten_metrics::monitored_mpsc::UnboundedReceiver;
 use parking_lot::Mutex;
 use prometheus::Registry;
-use std::{sync::Arc, time::Duration};
 use sui_protocol_config::{ConsensusNetwork, ProtocolConfig};
 use tempfile::TempDir;
-use tracing::info;
 
+use crate::network::tonic_network::to_socket_addr;
 use crate::transaction::NoopTransactionVerifier;
 use crate::{
     CommitConsumer, CommitConsumerMonitor, CommittedSubDag, ConsensusAuthority, TransactionClient,
@@ -18,7 +25,7 @@ use crate::{
 
 #[derive(Clone)]
 #[allow(unused)]
-pub(crate) struct NodeConfig {
+pub(crate) struct Config {
     pub authority_index: AuthorityIndex,
     pub db_dir: Arc<TempDir>,
     pub committee: Committee,
@@ -28,19 +35,15 @@ pub(crate) struct NodeConfig {
     pub protocol_config: ProtocolConfig,
 }
 
-/// A wrapper struct to allow us manage the underlying AuthorityNode.
-#[allow(unused)]
-pub(crate) struct Node {
-    container: Mutex<Option<AuthorityNodeContainer>>,
-    config: NodeConfig,
+pub(crate) struct AuthorityNode {
+    inner: Mutex<Option<AuthorityNodeInner>>,
+    config: Config,
 }
 
-#[allow(unused)]
-impl Node {
-    /// Create a new Node from the provided `NodeConfig`.
-    pub fn new(config: NodeConfig) -> Self {
+impl AuthorityNode {
+    pub fn new(config: Config) -> Self {
         Self {
-            container: Default::default(),
+            inner: Default::default(),
             config,
         }
     }
@@ -54,16 +57,16 @@ impl Node {
     pub async fn start(&self) -> Result<()> {
         info!(index =% self.config.authority_index, "starting in-memory node");
         let config = self.config.clone();
-        *self.container.lock() = Some(AuthorityNodeContainer::spawn(config).await);
+        *self.inner.lock() = Some(AuthorityNodeInner::spawn(config).await);
         Ok(())
     }
 
     pub fn spawn_committed_subdag_consumer(&self) -> Result<()> {
         let authority_index = self.config.authority_index;
-        let container = self.container.lock();
-        if let Some(container) = container.as_ref() {
-            let mut commit_receiver = container.take_commit_receiver();
-            let commit_consumer_monitor = container.commit_consumer_monitor();
+        let inner = self.inner.lock();
+        if let Some(inner) = inner.as_ref() {
+            let mut commit_receiver = inner.take_commit_receiver();
+            let commit_consumer_monitor = inner.commit_consumer_monitor();
             let _handle = tokio::spawn(async move {
                 while let Some(subdag) = commit_receiver.recv().await {
                     info!(index =% authority_index, "received committed subdag");
@@ -75,47 +78,172 @@ impl Node {
     }
 
     pub fn commit_consumer_monitor(&self) -> Arc<CommitConsumerMonitor> {
-        let container = self.container.lock();
-        if let Some(container) = container.as_ref() {
-            container.commit_consumer_monitor()
+        let inner = self.inner.lock();
+        if let Some(inner) = inner.as_ref() {
+            inner.commit_consumer_monitor()
         } else {
-            panic!("Container not initialised");
+            panic!("Node not initialised");
         }
     }
 
     pub fn transaction_client(&self) -> Arc<TransactionClient> {
-        let container = self.container.lock();
-        if let Some(container) = container.as_ref() {
-            container.transaction_client()
+        let inner = self.inner.lock();
+        if let Some(inner) = inner.as_ref() {
+            inner.transaction_client()
         } else {
-            panic!("Container not initialised");
+            panic!("Node not initialised");
         }
     }
 
     /// Stop this Node
     pub fn stop(&self) {
         info!(index =% self.config.authority_index, "stopping in-memory node");
-        *self.container.lock() = None;
+        *self.inner.lock() = None;
         info!(index =% self.config.authority_index, "node stopped");
     }
 
     /// If this Node is currently running
     pub fn is_running(&self) -> bool {
-        self.container
-            .lock()
-            .as_ref()
-            .map_or(false, |c| c.is_alive())
+        self.inner.lock().as_ref().map_or(false, |c| c.is_alive())
+    }
+}
+
+pub(crate) struct AuthorityNodeInner {
+    handle: Option<NodeHandle>,
+    cancel_sender: Option<tokio::sync::watch::Sender<bool>>,
+    consensus_authority: ConsensusAuthority,
+    commit_receiver: ArcSwapOption<UnboundedReceiver<CommittedSubDag>>,
+    commit_consumer_monitor: Arc<CommitConsumerMonitor>,
+}
+
+#[derive(Debug)]
+struct NodeHandle {
+    node_id: sui_simulator::task::NodeId,
+}
+
+/// When dropped, stop and wait for the node running in this node to completely shutdown.
+impl Drop for AuthorityNodeInner {
+    fn drop(&mut self) {
+        if let Some(handle) = self.handle.take() {
+            tracing::info!("shutting down {}", handle.node_id);
+            sui_simulator::runtime::Handle::try_current().map(|h| h.delete_node(handle.node_id));
+        }
+    }
+}
+
+impl AuthorityNodeInner {
+    /// Spawn a new Node.
+    pub async fn spawn(config: Config) -> Self {
+        let (startup_sender, mut startup_receiver) = tokio::sync::watch::channel(false);
+        let (cancel_sender, cancel_receiver) = tokio::sync::watch::channel(false);
+
+        let handle = sui_simulator::runtime::Handle::current();
+        let builder = handle.create_node();
+
+        let authority = config.committee.authority(config.authority_index);
+        let socket_addr = to_socket_addr(&authority.address).unwrap();
+        let ip = match socket_addr {
+            SocketAddr::V4(v4) => IpAddr::V4(*v4.ip()),
+            _ => panic!("unsupported protocol"),
+        };
+        let init_receiver_swap = Arc::new(ArcSwapOption::empty());
+        let int_receiver_swap_clone = init_receiver_swap.clone();
+
+        let node = builder
+            .ip(ip)
+            .name(format!("{}", config.authority_index))
+            .init(move || {
+                info!("Node restarted");
+                let config = config.clone();
+                let mut cancel_receiver = cancel_receiver.clone();
+                let init_receiver_swap_clone = int_receiver_swap_clone.clone();
+                let startup_sender_clone = startup_sender.clone();
+
+                async move {
+                    let (consensus_authority, commit_receiver, commit_consumer_monitor) =
+                        super::node::make_authority(config).await;
+
+                    startup_sender_clone.send(true).ok();
+                    init_receiver_swap_clone.store(Some(Arc::new((
+                        consensus_authority,
+                        commit_receiver,
+                        commit_consumer_monitor,
+                    ))));
+
+                    // run until canceled
+                    loop {
+                        if cancel_receiver.changed().await.is_err() || *cancel_receiver.borrow() {
+                            break;
+                        }
+                    }
+                    trace!("cancellation received; shutting down thread");
+                }
+            })
+            .build();
+
+        startup_receiver.changed().await.unwrap();
+
+        let Some(init_tuple) = init_receiver_swap.swap(None) else {
+            panic!("Components should be initialised by now");
+        };
+
+        let Ok((consensus_authority, commit_receiver, commit_consumer_monitor)) =
+            Arc::try_unwrap(init_tuple)
+        else {
+            panic!("commit receiver still in use");
+        };
+
+        Self {
+            handle: Some(NodeHandle { node_id: node.id() }),
+            cancel_sender: Some(cancel_sender),
+            consensus_authority,
+            commit_receiver: ArcSwapOption::new(Some(Arc::new(commit_receiver))),
+            commit_consumer_monitor,
+        }
+    }
+
+    /// Check to see that the Node is still alive by checking if the receiving side of the
+    /// `cancel_sender` has been dropped.
+    pub fn is_alive(&self) -> bool {
+        if let Some(cancel_sender) = &self.cancel_sender {
+            // unless the node is deleted, it keeps a reference to its start up function, which
+            // keeps 1 receiver alive. If the node is actually running, the cloned receiver will
+            // also be alive, and receiver count will be 2.
+            cancel_sender.receiver_count() > 1
+        } else {
+            false
+        }
+    }
+
+    pub fn take_commit_receiver(&self) -> UnboundedReceiver<CommittedSubDag> {
+        if let Some(commit_receiver) = self.commit_receiver.swap(None) {
+            let Ok(commit_receiver) = Arc::try_unwrap(commit_receiver) else {
+                panic!("commit receiver still in use");
+            };
+
+            commit_receiver
+        } else {
+            panic!("commit receiver already taken");
+        }
+    }
+
+    pub fn commit_consumer_monitor(&self) -> Arc<CommitConsumerMonitor> {
+        self.commit_consumer_monitor.clone()
+    }
+
+    pub fn transaction_client(&self) -> Arc<crate::TransactionClient> {
+        self.consensus_authority.transaction_client()
     }
 }
 
 pub(crate) async fn make_authority(
-    config: NodeConfig,
+    config: Config,
 ) -> (
     ConsensusAuthority,
     UnboundedReceiver<CommittedSubDag>,
     Arc<CommitConsumerMonitor>,
 ) {
-    let NodeConfig {
+    let Config {
         authority_index,
         db_dir,
         committee,
